requirements.txt
fastapi
uvicorn
python-socketio[asyncio]
aiohttp